+++ skrub/tests/test_fuzzy_join.py
@@ -9,19 +9,33 @@ from pandas.testing import assert_frame_equal
 from sklearn.feature_extraction.text import HashingVectorizer
 
 from skrub import fuzzy_join
+from skrub._dataframe._polars import POLARS_SETUP
+from skrub._dataframe._test_utils import is_module_polars
+
+MODULES = [pd]
+ASSERT_TUPLES = [(pd, assert_frame_equal)]
+
+if POLARS_SETUP:
+    import polars as pl
+    from polars.testing import assert_frame_equal as assert_frame_equal_pl
+
+    MODULES.append(pl)
+    ASSERT_TUPLES.append((pl, assert_frame_equal_pl))
 
 
+@pytest.mark.parametrize("px", MODULES)
 @pytest.mark.parametrize(
     "analyzer",
     ["char", "char_wb", "word"],
 )
-def test_fuzzy_join(analyzer: Literal["char", "char_wb", "word"]) -> None:
+def test_fuzzy_join(px, analyzer: Literal["char", "char_wb", "word"]):
     """
     Testing if fuzzy_join results are as expected.
     """
-
-    df1 = pd.DataFrame({"a1": ["ana", "lala", "nana et sana", np.NaN]})
-    df2 = pd.DataFrame({"a2": ["anna", "lala et nana", "lana", "sana", np.NaN]})
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
+    df1 = px.DataFrame({"a1": ["ana", "lala", "nana et sana", np.NaN]})
+    df2 = px.DataFrame({"a2": ["anna", "lala et nana", "lana", "sana", np.NaN]})
 
     df_joined = fuzzy_join(
         left=df1,
@@ -53,19 +67,25 @@ def test_fuzzy_join(analyzer: Literal["char", "char_wb", "word"]) -> None:
     assert "a12" in df_on.columns
 
 
-def test_max_dist():
-    left = pd.DataFrame({"A": ["aa", "bb"]})
-    right = pd.DataFrame({"A": ["aa", "ba"], "B": [1, 2]})
+@pytest.mark.parametrize("px", MODULES)
+def test_max_dist(px):
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
+    left = px.DataFrame({"A": ["aa", "bb"]})
+    right = px.DataFrame({"A": ["aa", "ba"], "B": [1, 2]})
     join = fuzzy_join(left, right, on="A", suffix="r")
     assert join["Br"].to_list() == [1, 2]
     join = fuzzy_join(left, right, on="A", suffix="r", max_dist=0.5)
     assert join["Br"].fillna(-1).to_list() == [1, -1]
 
 
-def test_perfect_matches():
+@pytest.mark.parametrize("px", MODULES)
+def test_perfect_matches(px):
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
     # non-regression test for https://github.com/skrub-data/skrub/issues/764
     # fuzzy_join when all rows had a perfect match used to trigger a division by 0
-    df = pd.DataFrame({"A": [0, 1]})
+    df = px.DataFrame({"A": [0, 1]})
     with warnings.catch_warnings():
         warnings.simplefilter("error")
         warnings.filterwarnings("ignore", message="This feature is still experimental")
@@ -73,37 +93,46 @@ def test_perfect_matches():
     assert_array_equal(join["skrub.Joiner.rescaled_distance"].to_numpy(), [0.0, 0.0])
 
 
-def test_fuzzy_join_dtypes() -> None:
+@pytest.mark.parametrize("px", MODULES)
+def test_fuzzy_join_dtypes(px):
     """
     Test that the dtypes of dataframes are maintained after join
     """
-    a = pd.DataFrame({"col1": ["aaa", "bbb"], "col2": [1, 2]})
-    b = pd.DataFrame({"col1": ["aaa_", "bbb_"], "col3": [1, 2]})
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
+    a = px.DataFrame({"col1": ["aaa", "bbb"], "col2": [1, 2]})
+    b = px.DataFrame({"col1": ["aaa_", "bbb_"], "col3": [1, 2]})
     c = fuzzy_join(a, b, on="col1", suffix="r")
     assert a.dtypes["col2"].kind == "i"
     assert c.dtypes["col2"] == a.dtypes["col2"]
     assert c.dtypes["col3r"] == b.dtypes["col3"]
 
 
-def test_missing_keys() -> None:
-    a = pd.DataFrame({"col1": ["aaa", "bbb"], "col2": [1, 2]})
-    b = pd.DataFrame({"col1": ["aaa_", "bbb_"], "col3": [1, 2]})
+@pytest.mark.parametrize("px", MODULES)
+def test_missing_keys(px):
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
+    a = px.DataFrame({"col1": ["aaa", "bbb"], "col2": [1, 2]})
+    b = px.DataFrame({"col1": ["aaa_", "bbb_"], "col3": [1, 2]})
     with pytest.raises(
         ValueError,
         match=r"Must pass",
     ):
         fuzzy_join(a, b, left_on="col1", suffix="r", insert_match_info=False)
-    left = pd.DataFrame({"a": ["aa", np.NaN, "bb"], "b": [1, 2, np.NaN]})
-    right = pd.DataFrame(
+    left = px.DataFrame({"a": ["aa", np.NaN, "bb"], "b": [1, 2, np.NaN]})
+    right = px.DataFrame(
         {"a": ["aa", "bb", np.NaN, "cc", "dd"], "c": [5, 6, 7, 8, np.NaN]}
     )
     output = fuzzy_join(left, right, on="a", suffix="r", insert_match_info=False)
     assert output.shape == (3, 4)
 
 
-def test_drop_unmatched() -> None:
-    a = pd.DataFrame({"col1": ["aaaa", "bbb", "ddd dd"], "col2": [1, 2, 3]})
-    b = pd.DataFrame({"col1": ["aaa_", "bbb_", "cc ccc"], "col3": [1, 2, 3]})
+@pytest.mark.parametrize("px", MODULES)
+def test_drop_unmatched(px):
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
+    a = px.DataFrame({"col1": ["aaaa", "bbb", "ddd dd"], "col2": [1, 2, 3]})
+    b = px.DataFrame({"col1": ["aaa_", "bbb_", "cc ccc"], "col3": [1, 2, 3]})
 
     c1 = fuzzy_join(
         a,
@@ -115,12 +144,11 @@ def test_drop_unmatched() -> None:
         insert_match_info=False,
     )
     assert c1.shape == (2, 4)
-
     c2 = fuzzy_join(a, b, on="col1", max_dist=0.9, suffix="r", insert_match_info=False)
     assert sum(c2["col3r"].isna()) > 0
 
 
-def test_fuzzy_join_pandas_comparison() -> None:
+def test_fuzzy_join_pandas_comparison():
     """
     Tests if fuzzy_join's output is as similar as
     possible with `pandas.merge`.
@@ -146,13 +174,16 @@ def test_fuzzy_join_pandas_comparison() -> None:
         left, right, left_on="key", right_on="key_", insert_match_info=False
     )
 
-    pd.testing.assert_frame_equal(result, result_fj)
+    assert_frame_equal(result, result_fj)
 
 
-def test_correct_encoder() -> None:
+@pytest.mark.parametrize("px", MODULES)
+def test_correct_encoder(px):
     """
     Test that the encoder error checking is working as intended.
     """
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
 
     class TestVectorizer(HashingVectorizer):
         """
@@ -164,7 +195,7 @@ def test_correct_encoder() -> None:
         def fit(self, X, y=None):
             raise AssertionError("Custom vectorizer was called as intended.")
 
-    left = pd.DataFrame(
+    left = px.DataFrame(
         {
             "key": ["K0", "K1", "K2", "K3"],
             "A": ["A0", "A1", "A2", "A3"],
@@ -172,7 +203,7 @@ def test_correct_encoder() -> None:
         }
     )
 
-    right = pd.DataFrame(
+    right = px.DataFrame(
         {
             "key": ["K0", "K1", "K2", "K3"],
             "C": ["C0", "C1", "C2", "C3"],
@@ -188,13 +219,15 @@ def test_correct_encoder() -> None:
         fuzzy_join(left, right, on="key", string_encoder=enc, insert_match_info=False)
 
 
-def test_numerical_column() -> None:
+@pytest.mark.parametrize("px", MODULES)
+def test_numerical_column(px):
     """
     Testing that fuzzy_join works with numerical columns.
     """
-
-    left = pd.DataFrame({"str1": ["aa", "a", "bb"], "int": [10, 2, 5]})
-    right = pd.DataFrame(
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
+    left = px.DataFrame({"str1": ["aa", "a", "bb"], "int": [10, 2, 5]})
+    right = px.DataFrame(
         {
             "str2": ["aa", "bb", "a", "cc", "dd"],
             "int": [55, 6, 2, 15, 6],
@@ -223,21 +256,23 @@ def test_numerical_column() -> None:
     assert fj_num3.shape == (2, n_cols)
 
 
-def test_datetime_column():
+@pytest.mark.parametrize("px, assert_frame_equal_", ASSERT_TUPLES)
+def test_datetime_column(px, assert_frame_equal_):
     """
     Testing that fuzzy_join works with datetime columns.
     """
-
-    left = pd.DataFrame(
+    if is_module_polars(px):
+        pytest.xfail(reason="Module 'polars' has no attribute 'to_datetime'")
+    left = px.DataFrame(
         {
             "str1": ["aa", "a", "bb"],
-            "date": pd.to_datetime(["10/10/2022", "12/11/2021", "09/25/2011"]),
+            "date": px.to_datetime(["10/10/2022", "12/11/2021", "09/25/2011"]),
         }
     )
-    right = pd.DataFrame(
+    right = px.DataFrame(
         {
             "str2": ["aa", "bb", "a", "cc", "dd"],
-            "date": pd.to_datetime(
+            "date": px.to_datetime(
                 ["09/10/2022", "12/24/2021", "09/25/2010", "11/05/2025", "02/21/2000"]
             ),
         }
@@ -245,15 +280,15 @@ def test_datetime_column():
 
     fj_time = fuzzy_join(left, right, on="date", suffix="r", insert_match_info=False)
 
-    fj_time_expected = pd.DataFrame(
+    fj_time_expected = px.DataFrame(
         {
             "str1": ["aa", "a", "bb"],
-            "date": pd.to_datetime(["10/10/2022", "12/11/2021", "09/25/2011"]),
+            "date": px.to_datetime(["10/10/2022", "12/11/2021", "09/25/2011"]),
             "str2r": ["aa", "bb", "a"],
-            "dater": pd.to_datetime(["09/10/2022", "12/24/2021", "09/25/2010"]),
+            "dater": px.to_datetime(["09/10/2022", "12/24/2021", "09/25/2010"]),
         }
     )
-    assert_frame_equal(fj_time, fj_time_expected)
+    assert_frame_equal_(fj_time, fj_time_expected)
 
     n_cols = left.shape[1] + right.shape[1]
     n_samples = len(left)
@@ -275,27 +310,29 @@ def test_datetime_column():
     assert fj_time3.shape == (2, n_cols)
 
 
-def test_mixed_joins():
+@pytest.mark.parametrize("px, assert_frame_equal_", ASSERT_TUPLES)
+def test_mixed_joins(px, assert_frame_equal_):
     """
     Test fuzzy joining on mixed and multiple column types.
     """
-
-    left = pd.DataFrame(
+    if is_module_polars(px):
+        pytest.xfail(reason="Module 'polars' has no attribute 'to_datetime'")
+    left = px.DataFrame(
         {
             "str1": ["Paris", "Paris", "Paris"],
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": pd.to_datetime(["10/10/2022", "12/11/2021", "09/25/2011"]),
+            "date": px.to_datetime(["10/10/2022", "12/11/2021", "09/25/2011"]),
         }
     )
-    right = pd.DataFrame(
+    right = px.DataFrame(
         {
             "str_1": ["Paris", "Paris", "Paris", "cc", "dd"],
             "str_2": ["TX", "FR", "GR Mytho", "cc", "dd"],
             "int1": [55, 6, 2, 15, 6],
             "int2": [554, 146, 32, 215, 612],
-            "date": pd.to_datetime(
+            "date": px.to_datetime(
                 ["09/10/2022", "12/24/2021", "09/25/2010", "11/05/2025", "02/21/2000"]
             ),
         }
@@ -306,21 +343,21 @@ def test_mixed_joins():
         left, right, on=["int1", "int2"], suffix="r", insert_match_info=False
     )
 
-    expected_fj_num = pd.DataFrame(
+    expected_fj_num = px.DataFrame(
         {
             "str1": ["Paris", "Paris", "Paris"],
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": pd.to_datetime(["10/10/2022", "12/11/2021", "09/25/2011"]),
+            "date": px.to_datetime(["10/10/2022", "12/11/2021", "09/25/2011"]),
             "str_1r": ["Paris", "Paris", "dd"],
             "str_2r": ["FR", "FR", "dd"],
             "int1r": [6, 6, 6],
             "int2r": [146, 146, 612],
-            "dater": pd.to_datetime(["12/24/2021", "12/24/2021", "02/21/2000"]),
+            "dater": px.to_datetime(["12/24/2021", "12/24/2021", "02/21/2000"]),
         }
     )
-    assert_frame_equal(fj_num, expected_fj_num)
+    assert_frame_equal_(fj_num, expected_fj_num)
     assert fj_num.shape == (3, 10)
 
     # On multiple string keys
@@ -333,21 +370,21 @@ def test_mixed_joins():
         insert_match_info=False,
     )
 
-    expected_fj_str = pd.DataFrame(
+    expected_fj_str = px.DataFrame(
         {
             "str1": ["Paris", "Paris", "Paris"],
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": pd.to_datetime(["2022-10-10", "2021-12-11", "2011-09-25"]),
+            "date": px.to_datetime(["2022-10-10", "2021-12-11", "2011-09-25"]),
             "str_1r": ["Paris", "Paris", "Paris"],
             "str_2r": ["TX", "FR", "GR Mytho"],
             "int1r": [55, 6, 2],
             "int2r": [554, 146, 32],
-            "dater": pd.to_datetime(["2022-09-10", "2021-12-24", "2010-09-25"]),
+            "dater": px.to_datetime(["2022-09-10", "2021-12-24", "2010-09-25"]),
         }
     )
-    assert_frame_equal(fj_str, expected_fj_str)
+    assert_frame_equal_(fj_str, expected_fj_str)
     assert fj_str.shape == (3, 10)
 
     # On mixed, numeric and string keys
@@ -359,21 +396,21 @@ def test_mixed_joins():
         suffix="r",
         insert_match_info=False,
     )
-    expected_fj_mixed = pd.DataFrame(
+    expected_fj_mixed = px.DataFrame(
         {
             "str1": ["Paris", "Paris", "Paris"],
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": pd.to_datetime(["2022-10-10", "2021-12-11", "2011-09-25"]),
+            "date": px.to_datetime(["2022-10-10", "2021-12-11", "2011-09-25"]),
             "str_1r": ["Paris", "Paris", "Paris"],
             "str_2r": ["FR", "FR", "TX"],
             "int1r": [6, 6, 55],
             "int2r": [146, 146, 554],
-            "dater": pd.to_datetime(["2021-12-24", "2021-12-24", "2022-09-10"]),
+            "dater": px.to_datetime(["2021-12-24", "2021-12-24", "2022-09-10"]),
         }
     )
-    assert_frame_equal(fj_mixed, expected_fj_mixed)
+    assert_frame_equal_(fj_mixed, expected_fj_mixed)
     assert fj_mixed.shape == (3, 10)
 
     # On mixed time and string keys
@@ -385,21 +422,21 @@ def test_mixed_joins():
         suffix="r",
         insert_match_info=False,
     )
-    expected_fj_mixed2 = pd.DataFrame(
+    expected_fj_mixed2 = px.DataFrame(
         {
             "str1": ["Paris", "Paris", "Paris"],
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": pd.to_datetime(["2022-10-10", "2021-12-11", "2011-09-25"]),
+            "date": px.to_datetime(["2022-10-10", "2021-12-11", "2011-09-25"]),
             "str_1r": ["Paris", "Paris", "Paris"],
             "str_2r": ["TX", "FR", "GR Mytho"],
             "int1r": [55, 6, 2],
             "int2r": [554, 146, 32],
-            "dater": pd.to_datetime(["2022-09-10", "2021-12-24", "2010-09-25"]),
+            "dater": px.to_datetime(["2022-09-10", "2021-12-24", "2010-09-25"]),
         }
     )
-    assert_frame_equal(fj_mixed2, expected_fj_mixed2)
+    assert_frame_equal_(fj_mixed2, expected_fj_mixed2)
     assert fj_mixed2.shape == (3, 10)
 
     # On mixed time and numbers keys
@@ -411,32 +448,35 @@ def test_mixed_joins():
         suffix="r",
         insert_match_info=False,
     )
-    expected_fj_mixed3 = pd.DataFrame(
+    expected_fj_mixed3 = px.DataFrame(
         {
             "str1": ["Paris", "Paris", "Paris"],
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": pd.to_datetime(["2022-10-10", "2021-12-11", "2011-09-25"]),
+            "date": px.to_datetime(["2022-10-10", "2021-12-11", "2011-09-25"]),
             "str_1r": ["Paris", "Paris", "Paris"],
             "str_2r": ["FR", "FR", "GR Mytho"],
             "int1r": [6, 6, 2],
             "int2r": [146, 146, 32],
-            "dater": pd.to_datetime(["2021-12-24", "2021-12-24", "2010-09-25"]),
+            "dater": px.to_datetime(["2021-12-24", "2021-12-24", "2010-09-25"]),
         }
     )
-    assert_frame_equal(fj_mixed3, expected_fj_mixed3)
+    assert_frame_equal_(fj_mixed3, expected_fj_mixed3)
     assert fj_mixed3.shape == (3, 10)
 
 
-def test_iterable_input() -> None:
+@pytest.mark.parametrize("px", MODULES)
+def test_iterable_input(px):
     """
     Test if iterable input: list, set, dictionary or tuple works.
     """
-    df1 = pd.DataFrame(
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
+    df1 = px.DataFrame(
         {"a": ["ana", "lala", "nana"], "str2": ["Texas", "France", "Greek God"]}
     )
-    df2 = pd.DataFrame(
+    df2 = px.DataFrame(
         {
             "a": ["anna", "lala", "ana", "nnana"],
             "str_2": ["TX", "FR", "GR Mytho", "dd"],
@@ -468,14 +508,16 @@ def test_iterable_input() -> None:
     ).shape == (3, 4)
 
 
-# TODO
 @pytest.mark.xfail
-def test_missing_values() -> None:
+@pytest.mark.parametrize("px", MODULES)
+def test_missing_values(px):
     """
     Test fuzzy joining on missing values.
     """
-    a = pd.DataFrame({"col1": ["aaaa", "bbb", "ddd dd"], "col2": [1, 2, 3]})
-    b = pd.DataFrame({"col3": [np.NaN, "bbb", "ddd dd"], "col4": [1, 2, 3]})
+    if is_module_polars(px):
+        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'")
+    a = px.DataFrame({"col1": ["aaaa", "bbb", "ddd dd"], "col2": [1, 2, 3]})
+    b = px.DataFrame({"col3": [np.NaN, "bbb", "ddd dd"], "col4": [1, 2, 3]})
 
     with pytest.warns(UserWarning, match=r"merging on missing values"):
         c = fuzzy_join(a, b, left_on="col1", right_on="col3", insert_match_info=False)